Move contact details to footer
